Create Wrappers with null
Wrappers with null
Unlike primitives, a wrapper variable can be uninitialized and has a value of null. A primitive variable that is not initialized is automatically set to its default value (0 in the case of ints). Since wrappers can have null values, this allows us to test if it has been set before using it (unlike primitives). In this exercise, we see the difference in uninitialized primitives and wrapper objects.

Instructions

Declare a variable of type int called x but do not set its value.
Declare a variable of type Integer called y but do not set its value.
Display (print) the value of x even though you did not initialize its value.
Test if y is null.