Create POJO, wrappers and java.math
POJO, wrappers and java.math
Every great mission starts with a strong launch! In this exercise, you'll define a Rocket class and write the code to launch it. You can find the implementation of Rocket on lines 23-46 if you are interested.

Ready for liftoff?

Instructions

Create an instance of the Rocket POJO.
Set the Rocket thrust field to "7770000" as a BigDecimal using the corresponding setter.
Set the Boolean wrapper manned field to true using the corresponding setter.