Create Autoboxing and foreach
Autoboxing and foreach
In this exercise, you will explore autoboxing and looping of a List - specifically a LinkedList of prices. You will add some double to the LinkedList and allow autoboxing to convert them to Double, and use the "foreach" loop to get the average price of the prices in the LinkedList.

Instructions

Construct a new LinkedList of Doubles and set the prices variable to it.
Add a new price (a double) 9.65 to the end of the prices list. Autoboxing will automatically convert the double to a Double before adding it.
Add a new price (a double) 1.35 to the beginning of the prices list. Autoboxing will automatically convert the double to a Double before adding it.
Use a "foreach" loop to iterate through and sum all prices.