Create Working with Queue
Working with Queue
Queues collect and returns objects in the order they were added. Queues also have a capacity. Here you create a waiting list Queue (specifically from an ArrayBlockingQueue) and add names to it. You see what happens when trying to add more names than the capacity allows.

Instructions

Import ArrayBlockingQueue for use in the application.
Construct a new ArrayBlockingQueue of Strings with a capacity of 3 and set the waitList variable to it.
Add a new name, "Tarah", to waitList.
Remove the first name on the waitList.