Create Checked exception
Checked exception
Opposite of RuntimeException, all checked exceptions must have a try-catch block or the code will not compile, let alone run. In this exercise, you will see that you must add a try/catch around a checked exception or else the application will not compile and will not run.

Instructions

Run the sample code without any changes to see that the call to Class.forName() requires handling leading to a compile error (error: compilation failed) if it not handled.
Uncomment the try, catch, and message displayed in the catch block.
Now try to run the application again and see that the application will compile and run because the checked exception has been handled.